Atualização conta_pagar.php

+++ favorecido_crud.php
@@ -18,7 +18,6 @@ if (filter_input(INPUT_SERVER, "REQUEST_METHOD") === "POST") {
 
                 $registro = new stdClass();
                 $registro = json_decode($_POST['registro']);
-                validaDados($registro);
 
                 $sql = "insert into favorecido(nome, usuario_id) VALUES (?, ?) ";
                 $conexao = new PDO("mysql:host=" . SERVIDOR . ";dbname=" . BANCO, USUARIO, SENHA);
@@ -73,14 +72,34 @@ if (filter_input(INPUT_SERVER, "REQUEST_METHOD") === "POST") {
                 $registro = new stdClass();
                 $registro = json_decode($_POST["registro"]);
 
-                $sql = "delete from favorecido where id = ? ";
-                $conexao = new PDO("mysql:host=" . SERVIDOR . ";dbname=" . BANCO, USUARIO, SENHA);
-                $pre = $conexao->prepare($sql);
-                $pre->execute(array(
+                $select = new PDO("mysql:host=" . SERVIDOR . ";dbname=" . BANCO, USUARIO, SENHA);
+                $pro = $select->prepare('SELECT * FROM conta_receber WHERE favorecido = ?');
+                $pro->execute(array(
                     $registro->id
                 ));
 
-                print json_encode(1);
+                if (!$pro->rowCount() > 0) {
+                    $select = new PDO("mysql:host=" . SERVIDOR . ";dbname=" . BANCO, USUARIO, SENHA);
+                    $pra = $select->prepare('SELECT * FROM conta_pagar WHERE favorecido = ?');
+                    $pra->execute(array(
+                        $registro->id
+                    ));
+                    if(!$pra->rowCount() > 0){
+                        $sql = "delete from favorecido where id = ? ";
+                        $conexao = new PDO("mysql:host=" . SERVIDOR . ";dbname=" . BANCO, USUARIO, SENHA);
+                        $pre = $conexao->prepare($sql);
+                        $pre->execute(array(
+                            $registro->id
+                        ));
+
+                        print json_decode(1);
+                    }else{
+                        echo "Erro: Este favorecido está sendo utilizado em alguma conta sua em contas a pagar!! exclua sua conta a pagar antes de excluir ele. <br>";
+                    }
+                }
+                else{
+                    echo "Erro: Este favorecido está sendo utilizado em alguma conta sua em contas a receber!! exclua sua conta a receber antes de excluir ele. <br>";
+                }
             } catch (Exception $e) {
                 echo "Erro: " . $e->getMessage() . "<br>";
             } finally {
@@ -131,35 +150,12 @@ function buscarfavorecido(int $id)
     }
 }
 
-//consulta sem ajax
-function listarfavorecido()
+function listarFavorecido()
 {
     try {
         $usuario_id = isset($_SESSION["usuario_id"]) ? $_SESSION["usuario_id"] : 0;
 
         $sql = "select * from favorecido where usuario_id = ? order by nome";
-        $conexao = new PDO("mysql:host=" . SERVIDOR . ";dbname=" . BANCO, USUARIO, SENHA);
-        $pre = $conexao->prepare($sql);
-        $pre->execute(array(
-            $usuario_id
-        ));
-        $pre->execute();
-
-        return $pre->fetchAll(PDO::FETCH_ASSOC);
-    } catch (Exception $e) {
-        echo "Erro: " . $e->getMessage() . "<br>";
-    } finally {
-        $conexao = null;
-    }
-}
-
-//consulta sem ajax
-function listarfavorecidoEntrada()
-{
-    try {
-        $usuario_id = isset($_SESSION["usuario_id"]) ? $_SESSION["usuario_id"] : 0;
-
-        $sql = "select * from favorecido where usuario_id = ? and tipo = 1 order by nome";
         $conexao = new PDO("mysql:host=" . SERVIDOR . ";dbname=" . BANCO, USUARIO, SENHA);
         $pre = $conexao->prepare($sql);
         $pre->execute(array(
